Moved named pipe hosting files into Hosting.NamedPipes namespace
