+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Hosting/NamedPipes/INamedPipeProjectInfoDriver.cs
@@ -5,7 +5,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Razor.ProjectSystem;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Hosting;
+namespace Microsoft.AspNetCore.Razor.LanguageServer.Hosting.NamedPipes;
 
 internal interface INamedPipeProjectInfoDriver : IRazorProjectInfoDriver
 {
